backend/telemetry-client/src/requirements.txt
grpcio
grpcio-tools
PyYAML
psutil